// src/models/bifrost/BifrostFormQuestions/ReservedBifrostFormQuestionIds.ts
export enum ReservedBifrostFormQuestionIds {
  FIRST_NAME = "First Name",
  LAST_NAME = "Last Name",
  EMAIL = "Email",
  PHONE = "Phone",
  INQUIRY_DETAILS = "Inquiry Details",
  REASON_FOR_TRAVEL = "Reason for Travel",
  ARE_ITINERARY_DATES_FLEXIBLE = "Are Itinerary Dates Flexible",

  // Dates
  DATES = "Dates",
  POTENTIAL_DATES = "Potential Dates",
  FLEXIBLE_DATE_DESCRIPTION = "Flexible Date Description",

  ESTIMATED_GUEST_COUNT = "Estimated Guest Count",
  COUNT_OF_ROOMS_NEEDED = "Count of Rooms Needed",

  PAYMENT_SPLIT = "Payment Split",
}