// src/models/bifrost/BifrostFormQuestions/ReservedBifrostFormQuestionValues.ts
export enum ReservedBifrostReasonForTravelOptionValues {
  BUSINESS = "BUSINESS",
  SOCIAL = "SOCIAL",
  OTHER = "OTHER",
  EXTENDED_STAY = "EXTENDED_STAY",
}

export enum ReservedBifrostSplitPaymentOptionValues {
  SINGLE_PAYER = "SINGLE_PAYER",
  SPLIT_PAYER = "SPLIT_PAYER",
}

export enum ReservedBifrostDateFlexibilityOptionValues {
  FIRM_DATES = "FIRM_DATES",
  FLEXIBLE_DATES = "FLEXIBLE_DATES",
  STILL_DECIDING = "STILL_DECIDING",
}